First commit using compiled program instead of fish functions
Good thing we have git history, because I might want to revert.

- Move completely out of config.fish, and into conf.d/init.fish
- Only handle pre-exec and post-exec. Skip shell-exit for now. We weren't
getting anything useful out of it anyhow.
- Instead of everything implemented in fish (which worked very well, I have to say)
I'm experimenting with using a compiled binary that does all the work these
functions were doing before. It's an excuse to play around with Crystal.

TODOS:
- Move the binary into $PATH, and stop using the hardcoded
path to the development version of the file.
- I still may have to use some fish kung-fu after all: I want to implement
automatic backups once a day. I think the Crystal sqlite shard does not have
the backup api implemented, so I have to do it via the command-line shell.