Clashes with different versions of md5
gnu md5sum gets installed with coreutils, which
gets picked up by the `fish_md5` function.

Instead of piping into the function, just use it
like it was supposed to and direct the output of
the begin..end block to a file and pass it to
the `fish_md5` function.